requirements.txt
django
django_annoying
django_extensions
Werkzeug
south
django-tastypie
django-debug-toolbar

-e git+https://github.com/Shopify/shopify_python_api.git#egg=shopify

# Email template system
-e git+https://github.com/bradwhittington/django-templated-email.git#egg=django-templated-email

# Tasks
# apt-get install rabbitmq-server
django-celery

# Thumbnails
pil
-e git+https://github.com/sorl/sorl-thumbnail.git#egg=sorl-thumbnail
